main code

+++ 11614 Etruscan Warriors Never Play Chess.cpp
@@ -4,9 +4,9 @@ int main(){
     int t;
     cin>>t;
     while(t--){
-        int n;
+        long long int n;
         cin>>n;
-        int x=sqrt(1+(8*n));
+        long long int x=sqrt(1+(8*n));
         cout<<(x-1)/2<<endl;
     }
 }
